refactor: ControllerTest 중복 코드 제거
