+++ src/test/java/roomescape/controller/ControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.reservation;
+package roomescape.controller;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
